fix Recent() behavior when called with (now()-durationOfLevel, now())
The behavior before was to exclude the current bin from a level to have the complete level.duration() as history.
This made sense for the highest granularity (seconds) but for every other granularity this logic failed.
The 1 minute level for example could not be queried at current time. Now all levels have duration `end - length*granularity - 1*granularity`.
Which results in not being able to do `Recent(60*time.seconds)` using second granularity. Instead minute granularity will be used.
Increasing the length of the second granularity level from 60 to 61 would probably solve this."